requirements.txt
requests
numpy
opencv-python
pillow
pytest
allure-pytest
